got bar moving
