+++ src/App.css
@@ -1,5 +1,4 @@
 .App {
-  text-align: center;
 }
 
 .App-logo {
