+++ src/components/Movable/movable.css
@@ -1,0 +1,13 @@
+.movable {
+  background-color: forestgreen;
+  width: 100px;
+  height: 100px;
+  position: absolute;
+  bottom: 0;
+  transition: all 2s;
+}
+.script-box {
+  will-change: width, height, left, top;
+  position: relative;
+  width: 0;
+}
