+++ src/components/Visualizer/visualizer.css
@@ -2,4 +2,5 @@
   margin: 3rem;
   position: relative;
   height: 600px;
+  border: 1px solid black;
 }
